requirements.txt
﻿asgiref==3.8.1
distlib==0.3.9
dj-database-url==2.2.0
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.0
gunicorn